+++ system-tests/projects/vue3-webpack-ts-configured/cypress/support/component.ts
@@ -1,0 +1,3 @@
+import { mount } from 'cypress/vue'
+
+Cypress.Commands.add('mount', mount)