+++ system-tests/projects/vuecli5vue3-configured/cypress/support/component.js
@@ -1,1 +1,0 @@
-// TODO: source the example support file